Trivia component, answer choices
Created a Form in Trivia component (or a question componet)
That holds the input choices (answer choices)

When a choice is selected, we want to update our objects, that this choice is selected in this question component. Thus, we rerender(is that good? we dont need to, but we do need to update the props, so yeah)

NEXT:
- When we submit all our answers, gotta check which are right and wrong
- check-answer btn => Quiz component and render the proper UI.
- It will also, go to Trivia component and render the proper style.